+++ data/raw/f_307_indraneil.py
@@ -95,11 +95,10 @@ class TestCases(unittest.TestCase):
 
         with open(os.path.join(self.test_directory, "file2.json"), "w") as file:
             json.dump(self.json_data2, file)
-        super(TestCases, self).setUp()
 
     def tearDown(self):
-        shutil.rmtree(self.test_directory)
-        super(TestCases, self).tearDown()
+        if os.path.exists(self.test_directory):
+            shutil.rmtree(self.test_directory)
 
     def test_case_1(self):
         # Test with the sample directory created
